refactor: 데이터베이스 연결 예외 처리
