+++ nextstep-jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -1,5 +1,7 @@
 package nextstep.jdbc;
 
+import nextstep.jdbc.exception.DataAccessException;
+import nextstep.jdbc.exception.FailConnectionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -13,8 +15,17 @@ import java.util.List;
 public class JdbcTemplate {
     private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
 
+    private Connection getConnection() {
+        try {
+            return ConnectionManager.getConnection();
+        } catch (Exception e) {
+            log.error("Fail connection", e);
+            throw new FailConnectionException(e);
+        }
+    }
+
     public int update(String query, PreparedStatementSetter setter) {
-        try (Connection con = ConnectionManager.getConnection();
+        try (Connection con = getConnection();
              PreparedStatement pstmt = con.prepareStatement(query)) {
             setter.values(pstmt);
             return pstmt.executeUpdate();
@@ -29,7 +40,7 @@ public class JdbcTemplate {
     }
 
     public <T> List<T> query(String query, RowMapper<T> rowMapper, PreparedStatementSetter setter) {
-        try (Connection con = ConnectionManager.getConnection();
+        try (Connection con = getConnection();
              PreparedStatement pstmt = con.prepareStatement(query)) {
             setter.values(pstmt);
             return getQueryResults(rowMapper, pstmt);
@@ -58,7 +69,7 @@ public class JdbcTemplate {
     }
 
     public <T> T queryForObject(String query, RowMapper<T> rowMapper, PreparedStatementSetter setter) {
-        try (Connection con = ConnectionManager.getConnection();
+        try (Connection con = getConnection();
              PreparedStatement pstmt = con.prepareStatement(query)) {
             setter.values(pstmt);
             return getQueryResult(rowMapper, pstmt);
